move critics to src since no longer templated
