+++ include/mppic/critics/goal_angle_critic.hpp
@@ -14,19 +14,7 @@ class GoalAngleCritic : public CriticFunction
 {
 public:
 
-  void initialize() override
-  {
-    auto node = parent_.lock();
-    auto getParam = utils::getParamGetter(node, name_);
-    
-    getParam(power_, "goal_angle_cost_power", 1);
-    getParam(weight_, "goal_angle_cost_weight", 15);
-    getParam(threshold_to_consider_goal_angle_, "threshold_to_consider_goal_angle", 0.30);
-    RCLCPP_INFO(
-      logger_,
-      "GoalAngleCritic instantiated with %d power, %f weight, and %f angular threshold.",
-      power_, weight_, threshold_to_consider_goal_angle_);
-  }
+  void initialize() override;
 
   /**
    * @brief Evaluate cost related to robot orientation at goal pose
@@ -36,22 +24,7 @@ public:
    */
   virtual void score(
     const geometry_msgs::msg::PoseStamped & robot_pose, const xt::xtensor<double, 3> & trajectories,
-    const xt::xtensor<double, 2> & path, xt::xtensor<double, 1> & costs) override
-  {
-    xt::xtensor<double, 1> tensor_pose = {
-      static_cast<double>(robot_pose.pose.position.x), static_cast<double>(robot_pose.pose.position.y)};
-
-    auto path_points = xt::view(path, -1, xt::range(0, 2));
-
-    double points_to_goal_dists = xt::norm_l2(tensor_pose - path_points, {0})();
-
-    if (points_to_goal_dists < threshold_to_consider_goal_angle_) {
-      auto yaws = xt::view(trajectories, xt::all(), xt::all(), 2);
-      auto goal_yaw = xt::view(path, -1, 2);
-
-      costs += xt::pow(xt::mean(xt::abs(yaws - goal_yaw), {1}) * weight_, power_);
-    }
-  }
+    const xt::xtensor<double, 2> & path, xt::xtensor<double, 1> & costs) override;
 
 protected:
   double threshold_to_consider_goal_angle_{0};
